requirements.txt
psycopg2-binary==2.7.6.1
pytest==4.0.2
requests==2.21.0
